+++ engine/apps/schedules/tests/test_shift_swap_request.py
@@ -2,10 +2,14 @@ import datetime
 from unittest.mock import patch
 
 import pytest
+import pytz
 from django.utils import timezone
 
 from apps.schedules import exceptions
 from apps.schedules.models import CustomOnCallShift, ShiftSwapRequest
+from apps.user_management.models import User
+
+ROTATION_START = datetime.datetime(2150, 8, 29, 0, 0, 0, 0, tzinfo=pytz.UTC)
 
 
 @pytest.mark.django_db
@@ -119,7 +123,9 @@ def test_take_own_ssr(shift_swap_request_setup) -> None:
         ssr.take(beneficiary)
 
 
-@pytest.mark.skip(reason="Skipping to unblock release")
+@pytest.mark.skip(
+    "Skipping as flaky based on time of day that the test runs. Example failure here https://github.com/grafana/oncall/actions/runs/5747168275/job/15577755519?pr=2725#step:5:1005"
+)
 @pytest.mark.django_db
 def test_related_shifts(shift_swap_request_setup, make_on_call_shift) -> None:
     ssr, beneficiary, _ = shift_swap_request_setup()
@@ -152,3 +158,12 @@ def test_related_shifts(shift_swap_request_setup, make_on_call_shift) -> None:
     ]
     returned_events = [(e["start"], e["end"], e["users"][0]["pk"], e["users"][0]["swap_request"]["pk"]) for e in events]
     assert returned_events == expected
+
+
+@pytest.mark.django_db
+def test_possible_benefactors(shift_swap_request_setup) -> None:
+    ssr, beneficiary, benefactor = shift_swap_request_setup()
+
+    with patch.object(ssr.schedule, "related_users") as mock_related_users:
+        mock_related_users.return_value = User.objects.filter(pk__in=[beneficiary.pk, benefactor.pk])
+        assert list(ssr.possible_benefactors) == [benefactor]