+++ engine/apps/slack/types/interaction_payloads/__init__.py
@@ -5,20 +5,11 @@ from .shortcuts import MessageActionPayload
 from .slash_command import SlashCommandPayload
 from .view_submission import ViewSubmissionPayload
 
-
-class EventPayload:
-    BlockActionsPayload = BlockActionsPayload
-    DialogSubmissionPayload = DialogSubmissionPayload
-    InteractiveMessagesPayload = InteractiveMessagesPayload
-    MessageActionPayload = MessageActionPayload
-    SlashCommandPayload = SlashCommandPayload
-    ViewSubmissionPayload = ViewSubmissionPayload
-
-    Any = (
-        BlockActionsPayload
-        | DialogSubmissionPayload
-        | InteractiveMessagesPayload
-        | MessageActionPayload
-        | SlashCommandPayload
-        | ViewSubmissionPayload
-    )
+EventPayload = (
+    BlockActionsPayload
+    | DialogSubmissionPayload
+    | InteractiveMessagesPayload
+    | MessageActionPayload
+    | SlashCommandPayload
+    | ViewSubmissionPayload
+)